Add invalidate
